+++ api/src/test/java/org/btrplace/plan/DiscreteViolationExceptionTest.java
@@ -21,40 +21,22 @@ package org.btrplace.plan;
 import org.btrplace.model.DefaultModel;
 import org.btrplace.model.Model;
 import org.btrplace.model.constraint.SatConstraint;
-import org.btrplace.plan.event.Action;
 import org.mockito.Mockito;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
 /**
- * Basic unit tests for {@link ReconfigurationPlanCheckerException}.
- *
  * @author Fabien Hermenier
  */
-public class ReconfigurationPlanCheckerExceptionTest {
+public class DiscreteViolationExceptionTest {
 
     @Test
-    public void testInstantiationWithModel() {
+    public void test() {
         SatConstraint c = Mockito.mock(SatConstraint.class);
         Model m = new DefaultModel();
-        ReconfigurationPlanCheckerException ex = new ReconfigurationPlanCheckerException(c, m, true);
+        DiscreteViolationException ex = new DiscreteViolationException(c, m);
         Assert.assertEquals(ex.getModel(), m);
-        Assert.assertEquals(ex.isOrigin(), true);
-        Assert.assertNull(ex.getAction());
-
-        ex = new ReconfigurationPlanCheckerException(c, m, false);
-        Assert.assertEquals(ex.isOrigin(), false);
         Assert.assertEquals(ex.getConstraint(), c);
-        Assert.assertFalse(ex.toString().contains("null"));
-    }
-
-    @Test
-    public void testInstantiationWithAction() {
-        SatConstraint c = Mockito.mock(SatConstraint.class);
-        Action a = Mockito.mock(Action.class);
-        ReconfigurationPlanCheckerException ex = new ReconfigurationPlanCheckerException(c, a);
-        Assert.assertEquals(ex.getAction(), a);
-        Assert.assertNull(ex.getModel());
         Assert.assertFalse(ex.toString().contains("null"));
     }
 }